Simple_Canvas/requirements.txt
numpy==1.18.0
opencv-contrib-python==3.4.1.15
pandas==0.25.3
Pillow==7.0.0
PyQt5==5.14.0
PyQt5-sip==12.7.0
pyserial==3.4
pytesseract==0.3.1
python-dateutil==2.8.1
pytz==2019.3
qimage2ndarray==1.8.2
scipy==1.4.1
six==1.13.0
